+++ Render_Engine/Util/Point3D.cs
@@ -6,9 +6,9 @@
     /// </summary>
     internal class Point3D
     {
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
 
         /// <summary>
         /// Default constructor.
@@ -27,7 +27,7 @@
         /// <param name="x">X coordinate.</param>
         /// <param name="y">Y coordinate.</param>
         /// <param name="z">Z coordinate.</param>
-        public Point3D(float x, float y, float z)
+        public Point3D(double x, double y, double z)
         {
             X = x;
             Y = y;
@@ -44,7 +44,7 @@
         /// <summary>
         /// Returns the largest of the three coordinates.
         /// </summary>
-        public float Max()
+        public double Max()
         {
             if (X > Y && X > Z) return X;
             if (Y > Z) return Y;
@@ -54,7 +54,7 @@
         /// <summary>
         /// Returns the smallest of the three coordinates.
         /// </summary>
-        public float Min()
+        public double Min()
         {
             if (X < Y && X < Z) return X;
             if (Y < Z) return Y;
@@ -67,7 +67,7 @@
         /// <param name="p1">First point.</param>
         /// <param name="p2">Second point.</param>
         /// <returns>Distance between <paramref name="p1"/> and <paramref name="p2"/>.</returns>
-        public static float distance(Point3D p1, Point3D p2) => (float)Math.Sqrt(Math.Pow(p1.X - p2.X, 2) + Math.Pow(p1.Y - p2.Y, 2) + Math.Pow(p1.Z - p2.Z, 2));
+        public static double distance(Point3D p1, Point3D p2) => (double)Math.Sqrt(Math.Pow(p1.X - p2.X, 2) + Math.Pow(p1.Y - p2.Y, 2) + Math.Pow(p1.Z - p2.Z, 2));
 
         /// <summary>
         /// Converts this point to a <see cref="Vector3D"/>.
@@ -81,9 +81,9 @@
         public static Point3D operator +(Point3D p, VectorClass v) => new Point3D(p.X + v.X, p.Y + v.Y, p.Z + v.Z);
         public static Point3D operator +(VectorClass v, Point3D p) => p + v;
         public static Vector3D operator -(Point3D a, Point3D b) => new Vector3D(a.X - b.X, a.Y - b.Y, a.Z - b.Z);
-        public static Point3D operator *(float scalair, Point3D p) => new Point3D(p.X * scalair, p.Y * scalair, p.Z * scalair);
-        public static Point3D operator *(Point3D p, float scalair) => scalair * p;
-        public static Point3D operator /(Point3D p, float scalair) => 1 / scalair * p;
+        public static Point3D operator *(double scalair, Point3D p) => new Point3D(p.X * scalair, p.Y * scalair, p.Z * scalair);
+        public static Point3D operator *(Point3D p, double scalair) => scalair * p;
+        public static Point3D operator /(Point3D p, double scalair) => 1 / scalair * p;
 
         public static bool operator ==(Point3D a, Point3D b) => a.X == b.X && a.Y == b.Y && a.Z == b.Z;
         public static bool operator !=(Point3D a, Point3D b) => !(a == b);
